requirements.txt
curl_cffi
beautifulsoup4
lxml
python-dotenv
